Refactor models.py into separate files in models folder
