+++ models/myapp/models/__init__.py
@@ -1,0 +1,4 @@
+from .author import Author
+from .profile import Profile
+from .blog import Blog
+from .entry import Entry
