+++ models/myapp/models/author.py
@@ -1,0 +1,5 @@
+from django.db import models
+
+class Author(models.Model):
+    name = models.CharField(max_length=200)
+    email = models.EmailField()
