+++ models/myapp/models/blog.py
@@ -1,0 +1,4 @@
+from django.db import models
+
+class Blog(models.Model):
+    title = models.CharField(max_length=100)
