+++ models/myapp/models/entry.py
@@ -1,0 +1,9 @@
+from django.db import models
+from .blog import Blog
+from .author import Author
+
+class Entry(models.Model):
+    blog = models.ForeignKey(Blog, on_delete=models.CASCADE)
+    headline = models.CharField(max_length=255)
+    body_text = models.TextField()
+    authors = models.ManyToManyField(Author)
