+++ models/myapp/models/profile.py
@@ -1,0 +1,6 @@
+from django.db import models
+from .author import Author
+
+class Profile(models.Model):
+    user = models.OneToOneField(Author, on_delete=models.CASCADE)
+    bio = models.TextField()
